// backend/src/models/subject.model.ts
import mongoose,{Model,Schema,Document,Types} from "mongoose";


export interface ISubject extends Document{
    userId:Types.ObjectId
    title:string
    isAutoGenerated:boolean
}

const subjectSchema=new Schema<ISubject>({
    userId:{
        type:Schema.Types.ObjectId,
        ref:"User",
        required:true
    },
    title:{
        type:String,
        required:true,
        trim:true
    },
    isAutoGenerated:{
        type:Boolean,
        default:false
    }
},{timestamps:true})

subjectSchema.index({
    userId:1,
    title:1
})

subjectSchema.set('toJSON', {
  transform: (_doc, ret:any) => {
    delete ret.__v

    return ret
  },
})

const Subject:Model<ISubject>=mongoose.model<ISubject>("Subject",subjectSchema)

export default Subject